feat: 调整二维码生成接口.
